requirements.txt
openai>=1.80.0
python-dotenv==1.0.0 